EVENT_PLANNER_CHATBOT/requirements.txt
google-generativeai
openai
requests
flask
pydantic
python-dotenv
numpy
scikit-learn
colorama
pymongo
PyJWT
